* the DEFLATE uncompressed-block estimate is taken in 64-bit
3 + numSubBlocks * 5 * 8 + dataArray.Length * 8 would wrap for a 2^28-byte block, and
a negative estimate would make an uncompressed block look cheaper than any coded one.

It is not reachable today and was verified not to be: Write and Finish never hand
EmitCompressedBlock more than DefaultBlockSize (32768 bytes, 131072 at Maximum), so
the product tops out near one million. Measured, BB_Deflate compresses 268,435,455
and 268,435,457 bytes of text to the same 0.85 percent ratio with no step at the
boundary.

Widened anyway so that changing the block size cannot reintroduce the hazard, with
the actual bound written down next to it.

+++ Compression.Core/Deflate/DeflateCompressor.cs
@@ -157,9 +157,14 @@ public sealed class DeflateCompressor {
 
     litLenFreqs[DeflateConstants.EndOfBlock] = 1; // EOB
 
-    // Estimate uncompressed block cost: 3 header bits + 5-byte per sub-block header + raw bytes
+    // Estimate uncompressed block cost: 3 header bits + 5-byte per sub-block header + raw bytes.
+    // Taken in 64-bit. A 32-bit product would wrap for a 2^28-byte block, and the
+    // wrapped negative estimate would make an uncompressed block look cheaper than
+    // any Huffman-coded one. Write and Finish never hand this method more than
+    // DefaultBlockSize (32768, or 131072 at Maximum level) bytes, so the wrap is
+    // not reachable today; the width keeps it that way if the block size changes.
     var numSubBlocks = Math.Max(1, (dataArray.Length + DeflateCompressor.MaxBlockSize - 1) / DeflateCompressor.MaxBlockSize);
-    var uncompressedBits = 3 + numSubBlocks * 5 * 8 + dataArray.Length * 8;
+    var uncompressedBits = 3L + (long)numSubBlocks * 5 * 8 + (long)dataArray.Length * 8;
 
     if (this._level == DeflateCompressionLevel.Fast) {
       // Compare static Huffman vs uncompressed
